Lane-Follow-Train/requirements.txt
torch
torchvision
tqdm
pyyaml
onnx
opencv-python
numpy
albumentations